requirements.txt
fastapi
uvicorn
langchain-community
langchain-text-splitters
unstructured
python-magic
httpx
python-dotenv
aiokafka